+++ __tests__/provider/provider.notReady.js
@@ -1,0 +1,25 @@
+/* jshint -W117 */
+(function () {
+	'use strict';
+
+	const path    = require('path');
+	const helpers = require('yeoman-test');
+
+	jest.setTimeout(require('../../package.json').jestJasmineTimeout);
+
+	describe('generator-cozen-angular:provider', () => {
+		beforeAll(() => {
+			return helpers.run(path.join(__dirname, '../../generators/provider'))
+				.withPrompts({
+					areYouReady: false
+				})
+				.on('error', function ($error) {
+					console.log('Error !', $error);
+				});
+		});
+
+		it('is not ready');
+
+		it('close the current running generator');
+	});
+})();